Fix ImportWizardControllerTest.
In 1.x the existing authority entities were created using new MapEntity("Entity1")
This *looked* like it set the value of the ID attribute but in fact it set the *name* of the ID attribute.
The tests then tested that the authorities got added, but they should test that the authorities get updated for existing entities.