Bug fix; Updating the mail settings was broken.
If a user try'ed to update the mail settings, the update would fail and a error message would be shown indicating the settings could not be update due the the fact the Dynamic Entity could not
be converted to a MailSettingsImpl.

This update fixes the bug by using a (Dynamic)Entity in the decorators and creating new a
MailSettingsImpl from the (Dynamic)Entity when validating the update.

This commit may be viewed as a workaround, in future we need to design a more coherent way of updating system settings.