Fix #6747 When typing " in navigator I get error: {}
When the query in identified as invalid, no call to the backend should be made.
The user is shown an error explaining the query is invalid.

bug fix; move api inside of try + update test.

Would be good to fix this in backend someday or write a more complete validator as part of the api-client