chore(yarn): Use workspaces for the yarn modules
This will hoist shared libraries to a top-level node_modules dir.
A single run of yarn install will fetch them all, without the need for a mutex during builds.

The postcss loader searches for the postcss config relative to the loaded css, which fails if the css has been hoisted to a higher level module.
Specified the build dir as the location to search for the postcss config in utils.js