feat: case-insensitive like
The LIKE operator usually is executed in postgres.
Change it to ILIKE unless the attribute is tagged as case-sensitive.

When sent to the index, it already is case-insensitive but runs a prefix phrase query
on the analyzed text. This is the best we can do with the current indices.

For attributes tagged as token we make it exact by using a wildcard search.