chore: run api tests using jdk 17
Remove maven scripts to run the services.
Elasticsearch cannot run in jdk 17, it needs to run with its own jdk, so creating it from maven no longer works.
Travis was the last ci to use this in the vm.
Dev flow uses docker-compose, jenkins and circle ci use docker for the services.
Simplify the properties setup to use docker properties.